+++ last_task/.vscode/configurationCache.log
@@ -1,0 +1,1 @@
+{"buildTargets":[".PHONY","createdb","dropdb","migratedown","migrateup","postgres"],"launchTargets":[],"customConfigurationProvider":{"workspaceBrowse":{"browsePath":[],"compilerArgs":[]},"fileIndex":[]}}
