--- last_task/client/go.mod
@@ -1,3 +1,0 @@
-module github.com/xruterx/golang/last_task/client
-
-go 1.16
