+++ last_task/go.mod
@@ -1,8 +1,9 @@
-module github.com/xruterx/golang/last_task/server
+module github.com/xruterx/golang/last_task
 
 go 1.16
 
 require (
+	github.com/joho/godotenv v1.3.0
 	github.com/lib/pq v1.10.2
 	github.com/pkg/errors v0.9.1
 )
