+++ last_task/server/db.env
@@ -1,0 +1,9 @@
+ID_KEY=3b2ea0ea927a17e540f2330bcafdaba3
+SERVER_PORT=8080
+DB_TYPE=postgres
+DB_PORT=5432
+DB_TABLE=requests
+DB_USER=root
+DB_PASSWORD=root
+DB_HOST=localhost
+WEATHER_HOST=https://api.openweathermap.org/data/2.5/weather?q=%s&appid=%s
